chore: moved login page to server rendering
